+++ example/src/Components/TabTwo.tsx
@@ -1,11 +1,9 @@
-import React, { FC, Fragment } from 'react'
-
-const TabTwo: FC<{}> = () => {
+const TabTwo = () => {
   return (
-    <Fragment>
-      <h1>Tab Two</h1>
-      <p>Tab Two Content</p>
-    </Fragment>
+    <p>
+      <strong>Tab Two Content </strong>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Perferendis sint illum
+      iusto nostrum cumque qui voluptas tenetur inventore ut quis?
+    </p>
   )
 }
 export default TabTwo