+++ rollup.config.js
@@ -1,10 +1,9 @@
-import typescript from '@rollup/plugin-typescript'
+import typescript from 'rollup-plugin-typescript2'
 import { nodeResolve } from '@rollup/plugin-node-resolve'
 import commonjs from '@rollup/plugin-commonjs'
 import postcss from 'rollup-plugin-postcss'
 import { terser } from 'rollup-plugin-terser'
 import pkg from './package.json'
-// import image from '@rollup/plugin-image'
 
 export default {
   input: 'src/index.tsx',