+++ CHANGELOG.md
@@ -2,3 +2,11 @@
 
 - Bug fixed
 - Edit Readme
+
+## 07/31/2021
+
+- Refactor to useReducer
+- Refactor to [bootstrap](https://getbootstrap.com/docs/5.0/components/navs-tabs/) style
+- Removed unnecessarily code
+- Added eslint with prettier config
+- Updated: dependencies